feat(pipelineRouter.shared): rename pipelineRouter.shared and update imports
